loosen pylint contraints, fix strategies circular import
