+++ pandera/backends/polars/components.py
@@ -1,19 +1,17 @@
 """Validation backend for polars components."""
 
-from collections import defaultdict
 from typing import List, Optional, cast
 
 import polars as pl
 
 from pandera.api.polars.components import Column
-from pandera.backends.base import BaseSchemaBackend, CoreCheckResult
+from pandera.backends.base import CoreCheckResult
 from pandera.backends.polars.base import PolarsSchemaBackend
 from pandera.error_handlers import SchemaErrorHandler
 from pandera.errors import (
     SchemaError,
     SchemaErrors,
     SchemaErrorReason,
-    FailureCaseMetadata,
 )
